refactor activity profiler (#1219)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1219

Before this diff, `CuptiActivityProfiler` had code for Cupti *and* Rocm. The functionality for one or the other was controlled through compile-time `#ifdef`s. This refactor:

1. Creates `GenericActivityProfiler` which implements the generic logic that both Cupt and Rocm need. `ActivityProfilerController` now holds a `unique_ptr<GenericActivityProfiler>`. This class also implements the CPU only case. We can potentially revisit if that should be a separate class later.
2. Modifies `CuptiActivityProfiler` to be Cupti only.
3. Creates `RocmActivityProfiler` that is Rocm only.
4. Determines which kind of profiler to compile code for, Cupti or Rocm, through the build.

All code changes are downstream of the above changes. There's no new functionality here, so despite the number of lines changed by very large, there's no new logic.

The principle we're applying is that **different devices should be supported through different classes**. This diff is not making that completely true, as classes that use either `CuptiActivityProfiler` or `RocmActivityProfiler` have their own `#ifdef`. This refactor, though, is on the path to applying this principle in more places.

Note: I'm proposing we go with the name `RocmActivityProfiler` and not `RocActivityProfiler`. We already have `RoctracerLogger` and `RoctracerActivityApi`, but I think those are misnomers. I think we should rename those to also use the prefix `Rocm`.

Reviewed By: sraikund16

Differential Revision: D90651509

fbshipit-source-id: 919199e4aa9ba5d8f7843e545b7d6bee8642d5cb
(cherry picked from commit 445bc4f53d96527297110d778ba727cb6aa52806)
