+++ libkineto/src/ActivityProfilerController.h
@@ -21,8 +21,7 @@
 #include "ActivityProfilerInterface.h"
 #include "ActivityTraceInterface.h"
 #include "ConfigLoader.h"
-#include "CuptiActivityApi.h"
-#include "CuptiActivityProfiler.h"
+#include "GenericActivityProfiler.h"
 #include "InvariantViolations.h"
 #include "LoggerCollector.h"
 
@@ -70,21 +69,13 @@ class ActivityProfilerController : public ConfigLoader::ConfigHandler {
   void step();
   std::unique_ptr<ActivityTraceInterface> stopTrace();
 
-  bool isActive() {
-    return profiler_->isActive();
-  }
+  bool isActive();
 
-  void transferCpuTrace(std::unique_ptr<libkineto::CpuTraceBuffer> cpuTrace) {
-    return profiler_->transferCpuTrace(std::move(cpuTrace));
-  }
+  void transferCpuTrace(std::unique_ptr<libkineto::CpuTraceBuffer> cpuTrace);
 
-  void recordThreadInfo() {
-    profiler_->recordThreadInfo();
-  }
+  void recordThreadInfo();
 
-  void addChildActivityProfiler(std::unique_ptr<IActivityProfiler> profiler) {
-    profiler_->addChildActivityProfiler(std::move(profiler));
-  }
+  void addChildActivityProfiler(std::unique_ptr<IActivityProfiler> profiler);
 
   void addMetadata(const std::string& key, const std::string& value);
 
@@ -94,19 +85,13 @@ class ActivityProfilerController : public ConfigLoader::ConfigHandler {
       const std::string& error,
       const std::string& group_profile_id = "");
 
-  void pushCorrelationId(uint64_t id) {
-    profiler_->pushCorrelationId(id);
-  }
-  void popCorrelationId() {
-    profiler_->popCorrelationId();
-  }
+  void pushCorrelationId(uint64_t id);
 
-  void pushUserCorrelationId(uint64_t id) {
-    profiler_->pushUserCorrelationId(id);
-  }
-  void popUserCorrelationId() {
-    profiler_->popUserCorrelationId();
-  }
+  void popCorrelationId();
+
+  void pushUserCorrelationId(uint64_t id);
+
+  void popUserCorrelationId();
 
  private:
   bool shouldActivateIterationConfig(int64_t currentIter);
@@ -119,7 +104,7 @@ class ActivityProfilerController : public ConfigLoader::ConfigHandler {
   std::unique_ptr<Config> asyncRequestConfig_;
   std::mutex asyncConfigLock_;
 
-  std::unique_ptr<CuptiActivityProfiler> profiler_;
+  std::unique_ptr<GenericActivityProfiler> profiler_;
   std::unique_ptr<ActivityLogger> logger_;
   std::shared_ptr<LoggerCollector> loggerCollectorFactory_;
   std::thread* profilerThreads_[ThreadType::THREAD_MAX_COUNT] = {nullptr};
